Move notepad to new package
