+++ .gitignore
@@ -1,5 +1,13 @@
-*
-!.gitignore
-!*.go
-!*.md
-!go.mod
+# inaries for programs and plugins
+*.exe
+*.exe~
+*.dll
+*.so
+*.dylib
+
+# Test binary, built with `go test -c`
+*.test
+
+# Output of the go coverage tool, specifically when used with LiteIDE
+*.out
+*.py
